Rework the test system
It is possible this qualifies as "integration testing", but its
main purpose is really to just make it so I don't have to keep
typing the same stuff over... and over... and over...  Which
actually discourages systematic testing when making a change that
might break things, anyway.

Unfortunately, this approach does have to be updated if you change
the default messages.